SITL/requirements.txt
dronekit
dronekit-sitl
pymavlink
google-generativeai
future
wxPython
MAVProxy
wexpect